Fix release script

+++ release.py
@@ -13,7 +13,7 @@ def run_cmd(cmd, dry_run=False):
         print(" ".join([shlex.quote(c) for c in cmd]))
         return ""
     else:
-        return subprocess.check_output(cmd)
+        return subprocess.check_output(cmd).decode("utf-8")
 
 
 def main():
@@ -86,16 +86,19 @@ def main():
     run_cmd(["git", "tag", f"v{next_version}"], cli_args.dry_run)
 
     changelog = run_cmd(["poetry", "run", "gitchangelog"], cli_args.dry_run)
-    with open("CHANGELOG.md", "w") as ch:
-        ch.write(changelog.decode("utf-8"))
+    if not cli_args.dry_run:
+        with open("CHANGELOG.md", "w") as ch:
+            ch.write(changelog)
 
     run_cmd(["git", "commit", "--amend", "--no-edit"], cli_args.dry_run)
     # repeat tag for changelog (forced)
     run_cmd(["git", "tag", "-f", f"v{next_version}"], cli_args.dry_run)
 
     if not cli_args.no_push:
-        run_cmd(["git", "push"], cli_args.dry_run)
-        run_cmd(["git", "push", "--tags"], cli_args.dry_run)
+        remotes = run_cmd(["git", "remote"]).rstrip()
+        for remote in remotes.split("\n"):
+            run_cmd(["git", "push", remote], cli_args.dry_run)
+            run_cmd(["git", "push", remote, "--tags"], cli_args.dry_run)
 
 
 if __name__ == "__main__":
